Title: 겹치지 않게 선분 고르기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250303/겹치지 않게 선분 고르기/select-segments-without-overlap.cpp
@@ -22,7 +22,7 @@ bool is_overlapped(int l1, int r1) {
 
 void f(int pos) {
     if (pos == n) {
-        answer = max(answer, selected_line.size());
+        answer = max(answer, (int) selected_line.size());
         return;
     }
